requirements.txt
aiohttp==3.11.12
cdp_langchain==0.0.13
fastapi==0.115.8
langchain==0.3.18
langchain_community==0.3.17
langchain_core==0.3.35
langgraph==0.2.72
orjson==3.10.15
pandas==2.2.3
pydantic==2.10.6
python-dotenv==1.0.1
pytz==2025.1
Requests==2.32.3
schedule==1.2.2
uvicorn==0.34.0
web3==7.8.0
